simple envelope

+++ include/audiomath.h
@@ -2,7 +2,7 @@
 #define ROB_AUDIO_MATH_H
 #include <math.h>
 #ifndef TAU
-#define TAU (double)(2.0f * M_PI)
+#define TAU (double)((double)2.0f * (double)M_PI)
 #endif
 #define SAMPLE_RATE 48000
 #define BLOCK_SIZE 512 * 8
